backend/requirements.txt
langchain
langchain-community
langchain-groq
python-dotenv
pdfplumber
python-docx
pytesseract
Pillow
tqdm
unstructured
fastapi
uvicorn
sentence-transformers